requirements.txt
aiohttp==3.9.1
aiogram==3.4.1
python-dotenv==1.0.1
